Flatten test folder package structure.
