+++ CardForm/src/test/java/com/braintreepayments/cardform/CountryCodeEditTextTest.java
@@ -1,6 +1,6 @@
-package com.braintreepayments.cardform.view;
+package com.braintreepayments.cardform;
 
-import com.braintreepayments.cardform.MobileNumberEditText;
+import com.braintreepayments.cardform.CountryCodeEditText;
 import com.braintreepayments.cardform.R;
 import com.braintreepayments.cardform.test.TestActivity;
 
@@ -16,21 +16,21 @@ import static junit.framework.Assert.assertFalse;
 import static junit.framework.Assert.assertTrue;
 
 @RunWith(RobolectricTestRunner.class)
-public class MobileNumberEditTextTest {
+public class CountryCodeEditTextTest {
 
-    private MobileNumberEditText mView;
+    private CountryCodeEditText mView;
 
     @Before
     public void setup() {
         mView = Robolectric.setupActivity(TestActivity.class)
-                .findViewById(R.id.bt_card_form_mobile_number);
+                .findViewById(R.id.bt_card_form_country_code);
     }
 
     @Test
-    public void getMobileNumber_returnsStrippedMobileNumber() {
-        mView.setText("(555) 555-5555");
+    public void getCountryCode_returnsStrippedCountryCode() {
+        mView.setText("+86");
 
-        assertEquals("5555555555", mView.getMobileNumber());
+        assertEquals("86", mView.getCountryCode());
     }
 
     @Test
@@ -39,14 +39,9 @@ public class MobileNumberEditTextTest {
     }
 
     @Test
-    public void invalidIfFewerThan8Characters() {
-        mView.setText("123");
-        assertFalse(mView.isValid());
-    }
+    public void validIfNotEmpty() {
+        mView.setText("+86");
 
-    @Test
-    public void validIfMoreThan8Characters() {
-        mView.setText("12341234");
         assertTrue(mView.isValid());
     }
 
@@ -59,6 +54,7 @@ public class MobileNumberEditTextTest {
 
     @Test
     public void getErrorMessage_returnsErrorMessageWhenEmpty() {
-        assertEquals(RuntimeEnvironment.application.getString(R.string.bt_mobile_number_required), mView.getErrorMessage());
+        assertEquals(RuntimeEnvironment.application.getString(R.string.bt_country_code_required),
+                mView.getErrorMessage());
     }
 }
