+++ CardForm/src/test/java/com/braintreepayments/cardform/DigitsOnlyFilterTest.java
@@ -1,4 +1,4 @@
-package com.braintreepayments.cardform.view;
+package com.braintreepayments.cardform;
 
 import android.text.SpannableString;
