+++ CardForm/src/test/java/com/braintreepayments/cardform/TestInitialValueCheckBox.java
@@ -1,4 +1,4 @@
-package com.braintreepayments.cardform.view;
+package com.braintreepayments.cardform;
 
 import android.os.Parcelable;
