+++ CardForm/src/test/java/com/braintreepayments/cardform/ViewUtilsTest.java
@@ -1,4 +1,4 @@
-package com.braintreepayments.cardform.utils;
+package com.braintreepayments.cardform;
 
 import android.app.Activity;
